feat: 完成后端核心功能, 包括 ADMIN 权限, 订单流和全局异常处理

+++ src/main/java/com/example/shopsite/dto/ProductCreationRequest.java
@@ -3,6 +3,9 @@ package com.example.shopsite.dto;
 import lombok.Data;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
+
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.Min;
 
 @Data
@@ -15,7 +18,7 @@ public class ProductCreationRequest {
     
     @NotNull(message = "价格不能为空")
     @Min(value = 0, message = "价格必须大于0")
-    private Double price;
+    private BigDecimal price;
     
     @NotNull(message = "库存不能为空")
     @Min(value = 0, message = "库存不能小于0")
